+++ src/assets/imgs/index.js
@@ -236,6 +236,8 @@ export { default as SignInSpring } from "./formImgs/Signin/SignInSpring.svg";
 export { default as SignInAutumm } from "./formImgs/Signin/SignInAutumm.svg";
 export { default as SignInSummer } from "./formImgs/Signin/SignInSummer.svg";
 export { default as SignInWinter } from "./formImgs/Signin/SignInWinter.svg";
+export { default as backIcon } from "./formImgs/Signin/backIcon.svg";
+export { default as plusIcon } from "./formImgs/Signin/plusIcon.svg";
 
 // ------------------------SignUp------------
 export { default as SignUpSpring } from "./formImgs/SignUp/SingUpSpring.svg";
@@ -243,4 +245,3 @@ export { default as SignUpAutumm } from "./formImgs//SignUp/SignUpAutumm.svg";
 export { default as SignUpSummer } from "./formImgs/SignUp/SignUpSummer.svg";
 export { default as SignUpWinter } from "./formImgs/SignUp/SignUpWinter.svg";
 export { default as passwordCheck } from "./formImgs/Signin/passwordCheck.svg";
-export { default as backIcon } from "./formImgs/Signin/backIcon.svg";
