+++ src/components/header/AuthenticationNavbar/AuthenMedium.js
@@ -280,9 +280,9 @@ const AuthenMedium = () => {
                 <div
                   className={`${
                     !state?.hamburgerMenu
-                      ? "hamburger flex flex-col ease-linear duration-500 overscroll-none"
-                      : "left-[-450px] ll:-left-[500px] sm:left-[800px] ease-linear duration-500"
-                  }  top-[62px] left-0 fixed right-0 w-full bg-white z-10 h-full overflow-y-scroll px-[15px] ll:flex-fit`}
+                      ? "hamburger flex flex-col ease-linear duration-500  z-10 overscroll-none"
+                      : "left-[-1050px] z-[-10]  ease-linear duration-500"
+                  }  top-[62px] left-0 fixed right-0 w-full bg-white h-full overflow-y-scroll px-[15px] ll:flex-fit`}
                 >
                   {/* Searching section */}
                   <div className="search flex items-center justify-between rounded-lg font-AeonikProMedium h-10 mt-3 mb-3 border border-searchBg ss:mt-3 md:hidden w-full">
@@ -471,7 +471,7 @@ const AuthenMedium = () => {
                     />
                   );
                 })}
-              </div>
+              </div> 
 
               {/* Weather seection */}
               <div className="w-12 h-12 md:w-[120px] md:h-11 bg-btnBgColor border border-searchBgColor rounded-lg  md:rounded-lg ml-2">
